+++ src/main/java/data/bambda/BambdaFilter.java
@@ -2,8 +2,7 @@ package data.bambda;
 
 import data.ItemFilter;
 
-public class BambdaFilter implements ItemFilter<Bambda>
-{
+public class BambdaFilter implements ItemFilter<Bambda> {
     @Override
     public boolean filter(Bambda item, String searchTerm) {
         return item.name().toLowerCase().contains(searchTerm) ||
